requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
et_xmlfile==2.0.0
fastapi==0.121.2
h11==0.16.0
idna==3.11
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.12.4
pydantic_core==2.41.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
Telethon==1.42.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
websockets==15.0.1
